requirements.txt
fastapi==0.115.*
uvicorn==0.30.*
requests==2.*
httpx==0.27.*
pydantic==2.*
python-dotenv==1.*
pynput==1.*
python-multipart==0.0.*
PyYAML==6.*
Pillow==10.*          # optional fallback for screenshot capture on Windows
pyautogui==0.9.*       # optional (screenshots on Windows/Linux)
pyperclip==1.9.*
psutil==6.*
